Refactor billiards geometry and enable cue camera rig
