+++ webapp/src/pages/Games/PoolRoyale.jsx
@@ -25,6 +25,14 @@ import {
   CUE_RACK_PALETTE
 } from '../../utils/createCueRackDisplay.js';
 import {
+  adjustCornerNotchDepth,
+  adjustSideNotchDepth,
+  centroidFromRing,
+  multiPolygonToShapes,
+  scaleMultiPolygon,
+  signedRingArea
+} from '../../components/billiards/geometry.js';
+import {
   WOOD_FINISH_PRESETS,
   WOOD_GRAIN_OPTIONS,
   WOOD_GRAIN_OPTIONS_BY_ID,
@@ -38,147 +46,6 @@ import {
 import { MobilePortraitCameraRig, rad } from './simpleOrbitCamera';
 
 const LEGACY_SRGB_ENCODING = Reflect.get(THREE, 'sRGBEncoding');
-
-function signedRingArea(ring) {
-  let area = 0;
-  for (let i = 0; i < ring.length - 1; i++) {
-    const [x1, y1] = ring[i];
-    const [x2, y2] = ring[i + 1];
-    area += x1 * y2 - x2 * y1;
-  }
-  return area * 0.5;
-}
-
-function multiPolygonToShapes(mp) {
-  const shapes = [];
-  if (!Array.isArray(mp)) return shapes;
-  mp.forEach((poly) => {
-    if (!Array.isArray(poly) || !poly.length) return;
-    const outerRing = poly[0];
-    if (!Array.isArray(outerRing) || outerRing.length < 4) return;
-    const outerPts = outerRing.slice(0, -1);
-    if (!outerPts.length) return;
-    const outerClockwise = signedRingArea(outerRing) < 0;
-    const orderedOuter = outerClockwise
-      ? outerPts.slice().reverse()
-      : outerPts;
-    const shape = new THREE.Shape();
-    shape.moveTo(orderedOuter[0][0], orderedOuter[0][1]);
-    for (let i = 1; i < orderedOuter.length; i++) {
-      shape.lineTo(orderedOuter[i][0], orderedOuter[i][1]);
-    }
-    shape.closePath();
-
-    for (let ringIndex = 1; ringIndex < poly.length; ringIndex++) {
-      const holeRing = poly[ringIndex];
-      if (!Array.isArray(holeRing) || holeRing.length < 4) continue;
-      const holePts = holeRing.slice(0, -1);
-      if (!holePts.length) continue;
-      const holeClockwise = signedRingArea(holeRing) < 0;
-      const orderedHole = holeClockwise
-        ? holePts
-        : holePts.slice().reverse();
-      const hole = new THREE.Path();
-      hole.moveTo(orderedHole[0][0], orderedHole[0][1]);
-      for (let i = 1; i < orderedHole.length; i++) {
-        hole.lineTo(orderedHole[i][0], orderedHole[i][1]);
-      }
-      hole.closePath();
-      shape.holes.push(hole);
-    }
-
-    shapes.push(shape);
-  });
-  return shapes;
-}
-
-function centroidFromRing(ring) {
-  if (!Array.isArray(ring) || !ring.length) {
-    return { x: 0, y: 0 };
-  }
-  let sumX = 0;
-  let sumY = 0;
-  let count = 0;
-  for (let i = 0; i < ring.length; i++) {
-    const pt = ring[i];
-    if (!Array.isArray(pt) || pt.length < 2) continue;
-    sumX += pt[0];
-    sumY += pt[1];
-    count++;
-  }
-  if (!count) {
-    return { x: 0, y: 0 };
-  }
-  return { x: sumX / count, y: sumY / count };
-}
-
-function scaleMultiPolygon(mp, scale) {
-  if (!Array.isArray(mp) || typeof scale !== 'number') {
-    return [];
-  }
-  const clampedScale = Math.max(0, scale);
-  return mp
-    .map((poly) => {
-      if (!Array.isArray(poly) || !poly.length) return null;
-      const outerRing = poly[0];
-      const centroid = centroidFromRing(outerRing);
-      const scaledPoly = poly
-        .map((ring) => {
-          if (!Array.isArray(ring) || !ring.length) return null;
-          return ring
-            .map((pt) => {
-              if (!Array.isArray(pt) || pt.length < 2) return null;
-              const dx = pt[0] - centroid.x;
-              const dy = pt[1] - centroid.y;
-              return [centroid.x + dx * clampedScale, centroid.y + dy * clampedScale];
-            })
-            .filter(Boolean);
-        })
-        .filter((ring) => Array.isArray(ring) && ring.length > 0);
-      if (!scaledPoly.length) return null;
-      return scaledPoly;
-    })
-    .filter((poly) => Array.isArray(poly) && poly.length > 0);
-}
-
-function adjustCornerNotchDepth(mp, centerZ, sz) {
-  if (!Array.isArray(mp) || !Number.isFinite(centerZ) || !Number.isFinite(sz)) {
-    return Array.isArray(mp) ? mp : [];
-  }
-  return mp.map((poly) =>
-    Array.isArray(poly)
-      ? poly.map((ring) =>
-          Array.isArray(ring)
-            ? ring.map((pt) => {
-                if (!Array.isArray(pt) || pt.length < 2) return pt;
-                const [x, z] = pt;
-                const deltaZ = z - centerZ;
-                const towardCenter = -sz * deltaZ;
-                if (towardCenter <= 0) return [x, z];
-                return [x, centerZ + deltaZ * CHROME_CORNER_NOTCH_CENTER_SCALE];
-              })
-            : ring
-        )
-      : poly
-  );
-}
-
-function adjustSideNotchDepth(mp) {
-  if (!Array.isArray(mp)) return Array.isArray(mp) ? mp : [];
-  return mp.map((poly) =>
-    Array.isArray(poly)
-      ? poly.map((ring) =>
-          Array.isArray(ring)
-            ? ring.map((pt) => {
-                if (!Array.isArray(pt) || pt.length < 2) return pt;
-                const [x, z] = pt;
-                return [x, z * CHROME_SIDE_NOTCH_DEPTH_SCALE];
-              })
-            : ring
-        )
-      : poly
-  );
-}
 
 const POCKET_VISUAL_EXPANSION = 1.05;
 const CHROME_CORNER_POCKET_RADIUS_SCALE = 1;
@@ -4033,7 +3900,12 @@ function Table3D(
       ]]];
       union = polygonClipping.union(union, fieldClip);
     }
-    const adjusted = adjustCornerNotchDepth(union, cz, sz);
+    const adjusted = adjustCornerNotchDepth(
+      union,
+      cz,
+      sz,
+      CHROME_CORNER_NOTCH_CENTER_SCALE
+    );
     if (CHROME_CORNER_NOTCH_EXPANSION_SCALE === 1) {
       return adjusted;
     }
@@ -4067,7 +3939,7 @@ function Table3D(
     );
 
     const union = polygonClipping.union(circle, throat);
-    return adjustSideNotchDepth(union);
+    return adjustSideNotchDepth(union, CHROME_SIDE_NOTCH_DEPTH_SCALE);
   };
 
   const chromePlates = new THREE.Group();
